MCUX-73176 [edgefast-shell] RW612- HCI VSC to read back BT Annex 100 cal data does not work
[Description]
HCI VSC to read back BT Annex 100 cal data does not work

[Root Cause]
According to core_v5.4 Vol4 partE 5.4.1 hci command packet, ocf occupies 10 bits, and the current code only defines int8_t, so the ocf value in command(HCI_CMD_MARVELL_READ_CAL_DATA_ANNEX_100: 3f 0100 64) overflows and causes the sending command to be incorrect.

[Fix]
According to core spec requirements, expand ocf to 10 bits

[Testing]
Test pass.

[Open Issues/Todo]
 None

[Reviewers]
Song Xue, Author You

[Port proposal]
No-general

+++ source/shell/shell_test_mode.h
@@ -23,8 +23,8 @@ struct bt_hci_le_config {
 } __packed;
 
 struct bt_hci_command {
-	uint8_t ogf;
-	uint8_t ocf;
+	uint16_t ogf;
+	uint16_t ocf;
 	uint16_t opcode;
 	uint8_t param_len;
 };
